too many problems ; - ;
My game is still playable, but there are quite a few bugs. The one I’ve
been trying to fix is the one related to the cake, but as seen in my
problematic code, it’s been giving me a hard time…a few other grammar
mistakes fixed, but that might be about it.